// src/types/request.d.ts
export type Method = MethodV2 | MethodV3;

export interface EpisodeList {
  list: EList;
}

export interface EpisodeId {
  id: number;
}

export interface EpisodeListWithId extends EpisodeList, EpisodeId {
  list: EList.FAVORITES | EList.IGNORED;
}

export type ShowSearchParams = {
  network: number;
  genre: number;
  country: string;
  year: number;
  watching: number;
  category: string;
  status: string;
  sort: string;
  query: string;
};

export type SearchParams = {
  search: SearchObjectOptions;
};

export type Rating = 1 | 2 | 3 | 4 | 5;

export interface WithId {
  id: number;
}

export interface WithShowId {
  showId: number;
}

export interface WithLoginParam {
  login: string;
}

export interface WithRatingParam<R = Rating> extends WithId {
  rating?: R;
}

export interface WithCountParam extends WithId {
  count: number;
}

export interface WithStatusParam<S> extends WithId {
  status: S;
}

export interface WithQueryParam {
  query: string;
}

export interface WithSearchParam<S> {
  search: S;
  page?: number;
  pageSize?: number;
}

export interface WithMovieStatusParam<S> {
  movieId: number;
  status: S;
}

export type Credentials = {
  client_id: string;
  client_secret: string;
  username: string;
  password: string;
  grant_type?: 'password';
};

export type DefaultParams = {
  jsonrpc: string;
  method: Method;
  params: Record<string, unknown>;
  id: number;
};

export type SearchObjectOptions = Partial<{
  query: string;
  wasted: number;
  year: number;
  gender: EGender;
}>;

export type QueryParams = Record<string, unknown>;

export interface QueryHandlerParams<P = QueryParams, M = Method> {
  defaultParams: DefaultParams;
  method: M;
  params: P;
  query: AxiosInstance;
  url?: string;
}

export interface IMyShowsList {
  listsShows<T>(list: EList): Promise<RpcResponse<T, { list: EList }>>;

  listsAddShow<T>(id: number): Promise<RpcResponse<T, { id: number }>>;

  listsRemoveShow<T>(id: number): Promise<RpcResponse<T, WithId>>;

  listsEpisodes<T>(list: EList): Promise<RpcResponse<T, { list: EList }>>;

  listsAddEpisode<T>(
    id: number,
    list: EList,
  ): Promise<
    RpcResponse<
      T,
      {
        list: EList.FAVORITES | EList.IGNORED;
        id: number;
      }
    >
  >;

  listsRemoveEpisode<T>(
    id: number,
    list: EList.FAVORITES | EList.IGNORED,
  ): Promise<
    RpcResponse<
      T,
      {
        list: EList.FAVORITES | EList.IGNORED;
        id: number;
      }
    >
  >;
}

export interface IMyShowsManageV3 {
  manageSetShowStatus<T>(
    id: number,
    status: EShowStatus,
  ): Promise<RpcResponse<T, WithStatusParam<EShowStatus>>>;

  manageSetMovieStatus<T>(
    id: number,
    status: EMovieStatus,
  ): Promise<RpcResponse<T, WithMovieStatusParam<EMovieStatus>>>;
}

export interface IMyShowsManage extends IMyShowsManageV3 {
  manageRateShow<T, R = Rating>(
    id: number,
    rating: R,
  ): Promise<RpcResponse<T, WithRatingParam<R>>>;

  manageCheckEpisode<T, R = Rating>(
    id: number,
    rating?: R,
  ): Promise<
    RpcResponse<
      T,
      {
        id: number;
        rating?: R;
      }
    >
  >;

  manageUnCheckEpisode<T, R = Rating>(
    id: number,
  ): Promise<RpcResponse<T, WithId>>;

  manageRateEpisode<T, R = Rating>(
    id: number,
    rating: Rating,
  ): Promise<RpcResponse<T, WithRatingParam<R>>>;

  manageRateEpisode<T, R = Rating>(
    id: number,
    rating: R,
  ): Promise<RpcResponse<T, WithRatingParam<R>>>;

  manageRateEpisodesBulk<T, R = Rating>(
    id: number,
    r1: R[],
    r2: R[],
    r3: R[],
    r4: R[],
    r5: R[],
  ): Promise<RpcResponse<T, WithId>>;

  manageSyncEpisodes<T>(
    id: number,
    episodeIds: number[],
  ): Promise<
    RpcResponse<
      T,
      {
        showId: number;
        episodeIds: number[];
      }
    >
  >;

  manageSyncEpisodesDelta<T>(
    id: number,
    checkedIds: number[],
    unCheckedIds: number[],
  ): Promise<
    RpcResponse<
      T,
      {
        showId: number;
        checkedIds: number[];
        unCheckedIds: number[];
      }
    >
  >;

  manageReWatchShow(id: number, count: number): Promise<RpcResponse>;
}

export interface IMyShowsProfile {
  profileGet<T>(login: string): Promise<RpcResponse<T, WithLoginParam>>;

  profileFeed<T>(login: string): Promise<RpcResponse<T, WithLoginParam>>;

  profileFriends<T>(login: string): Promise<RpcResponse<T, WithLoginParam>>;

  profileFollowers<T>(login: string): Promise<RpcResponse<T, WithLoginParam>>;

  profileFriendsFeed<T>(): Promise<RpcResponse<T>>;

  profileShows<T>(login: string): Promise<RpcResponse<T[], WithLoginParam>>;

  profileEpisodes<T>(showId: number): Promise<RpcResponse<T, WithShowId>>;

  profileAchievements<T>(): Promise<RpcResponse<T>>;

  profileNewComments<T>(): Promise<RpcResponse<T>>;
}

export interface IMyShowsMovies {
  moviesSearch<T>(query: string): Promise<RpcResponse<T>>;

  moviesGetById(id: number): Promise<RpcResponse<MovieModel>>;
}

export interface IMyShowsShows {
  showsGetById<T>(id: number, withEpisodes: boolean): Promise<RpcResponse<T>>;

  showsGetByExternalId<T>(
    id: number,
    source: EShowSources,
  ): Promise<RpcResponse<T>>;

  showsSearch<T>(query: string): Promise<RpcResponse<T>>;

  showsSearchByFile<T>(file: string): Promise<RpcResponse<T>>;

  showsIds<T>(fromId: number, count: number): Promise<RpcResponse<T>>;

  showsEpisode<T>(id: number): Promise<RpcResponse<T>>;

  showsGenres<T>(): Promise<RpcResponse<T>>;

  showsTop<T>(mode: EGenderVote, count: number): Promise<RpcResponse<T>>;

  showsViewEpisodeComments<T>(id: number): Promise<RpcResponse<T>>;

  showsTrackEpisodeComments<T>(
    id: number,
    isTracked: boolean,
  ): Promise<RpcResponse<T>>;

  showsVoteEpisodeComment<T>(
    id: number,
    isPositive: boolean,
  ): Promise<RpcResponse<T>>;

  showsPostEpisodeComment<T>(
    id: number,
    text: string,
    parentId: number,
  ): Promise<RpcResponse<T>>;

  showsTranslateEpisodeComment<T>(
    id: number,
    language: string,
  ): Promise<RpcResponse<T>>;
}

export interface IMyShowsUsers {
  usersSearch<T>(
    search: Record<string, unknown>,
    page: number,
    pageSize: number,
  ): Promise<RpcResponse<T>>;

  usersCount<T>(search: Record<string, unknown>): Promise<RpcResponse<T>>;

  usersFiltersCounters<T>(
    query: string,
  ): Promise<RpcResponse<T, WithSearchParam<WithQueryParam>>>;
}

export interface IMyShowsBase {
  defaultParams: DefaultParams;

  login(): Promise<RpcError | void>;

  loginV3(): Promise<RpcError | void>;

  generic<P, T = JSONRPCResponse>(
    method: Method,
    params: Record<string, unknown>,
  ): Promise<T | RpcError>;
}

export interface IMyShows
  extends IMyShowsUsers,
    IMyShowsProfile,
    IMyShowsList,
    IMyShowsManage,
    IMyShowsShows,
    IMyShowsMovies,
    IMyShowsBase {}